完成文件hash功能

+++ app/controllers/Post.java
@@ -7,6 +7,7 @@ import play.mvc.Controller;
 import play.mvc.Http;
 import play.mvc.Result;
 import utils.Constant;
+import utils.Crypt;
 
 import java.io.File;
 import java.io.IOException;
@@ -34,24 +35,34 @@ public class Post extends Controller {
             return badRequest();
         }
 
-        String fileName = pic.getFilename();
         // 检查MIME TYPE
         String contentType = pic.getContentType();
         if (!contentType.startsWith("image")) {
             return badRequest();
         }
+        // 获得文件后缀名
+        String fileName = pic.getFilename();
+        String fileExtension = fileName.substring(fileName.lastIndexOf('.'));
 
         File userDir = new File(Constant.UPLOAD_FILE_PATH + self.userId.toString()
                 + File.separator + "images" + File.separator);
 
         try {
+
             FileUtils.forceMkdir(userDir);
 
             File picFile = pic.getFile();
-            File picLocal = new File(userDir.getPath() + File.separator + fileName);
+            // 获得文件的HASH值作为文件名
+            // TODO Hash时间过长
+            String fileHash = Crypt.digestFile(picFile);
 
-            byte[] picBytes = IOUtils.toByteArray(FileUtils.openInputStream(picFile));
-            FileUtils.writeByteArrayToFile(picLocal, picBytes);
+            File picLocal = new File(userDir.getPath() + File.separator + fileHash + fileExtension);
+
+            // 如果上传的文件不存在
+            if (!picLocal.exists()) {
+                byte[] picBytes = IOUtils.toByteArray(FileUtils.openInputStream(picFile));
+                FileUtils.writeByteArrayToFile(picLocal, picBytes);
+            }
 
         } catch (IOException e) {
             e.printStackTrace();
